bond and angle Gamess inputs generation
The ability to generate Gamess inputs for bond and angle scans is
added. The Ginp.py input needs a “bond”, “angle” or “diva” keyword at
the end of the line. If there is none, “diva” is assumed by default.
(Numerous small improvements in _GeomCalc.py, GeomStr.py, IO.py and
Ginp.py.)

+++ ParFit/IO.py
@@ -22,7 +22,7 @@ def ginp_inp(input_fname):
     elif scan_type=="angl":
        t1,t2,t3=map(int,tup.split())    
        tup=(t1,t2,t3)
-    b,e,s=map(int,bes.split())
+    b,e,s=map(float,bes.split())
     bes=(b,e,s)
     return scan_type,gopt_scan_fnameb,tup,bes
